Do not clobber the 'statusline' and 'ruler' settings for proxybuffers.
By using autocommands and buftype=acwrite to implement the 'write' of a proxy
buffer, a temp file is no longer required.  Since the buffer is now in
memory, and doesn't require a tempname() backing file, a more meaningful file
name can be given to the buffer for display purposes.  With a meaningful file
name, it is no longer necessary to clobber the user's statusline or ruler.